backend-fastapi/requirements.txt
fastapi
uvicorn
requests
groq
pdfplumber
multipart
python-dotenv
pydantic
psycopg2-binary
supabase
scikit-learn
numpy
pandas